More campaign loader cleanup.
