+++ game/campaignloader/__init__.py
@@ -1,0 +1,1 @@
+from .campaign import Campaign
